# api/openapi.yaml
openapi: 3.1.2
info:
  title: API Management Tax
  version: 0.1.0
  summary: Governança tributária multi-jurisdicional para América Latina e Caribe
  description: >-
    Contrato inicial do produto interno. O nome da organização é configuração do ambiente;
    exemplos usam dados sintéticos e não representam aconselhamento jurídico ou tributário.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Mesmo host da aplicação
...
paths:
...
  /v1/indicators/office-scorecards:
    get:
      summary: Listar scorecards trimestrais dos escritórios
      tags: [Indicators]
      operationId: listOfficeScorecards
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Calcular IDE com os onze indicadores ponderados
      tags: [Indicators]
      operationId: createOfficeScorecard
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/indicators/etr-measurements:
    get:
      summary: Listar medições de ETR por país e trimestre
      tags: [Indicators]
      operationId: listEtrMeasurements
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Calcular ETR corrente, diferido, Pilar 2 e variações
      tags: [Indicators]
      operationId: createEtrMeasurement
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/indicators/demands:
    get:
      summary: Listar demandas e situação de SLA
      tags: [Indicators]
      operationId: listTaxDemands
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Registrar demanda e calcular prioridade e cumprimento de SLA
      tags: [Indicators]
      operationId: createTaxDemand
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/indicators/contingencies:
    get:
      summary: Listar contingências e respectivas exposições
      tags: [Indicators]
      operationId: listTaxContingencies
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Registrar contingência, aging, baseline e apetite de risco
      tags: [Indicators]
      operationId: createTaxContingency
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/indicators/executive-dashboard:
    get:
      summary: Consolidar IDE, ETR, SLA e risco para CFO e Board
      tags: [Indicators]
      operationId: getExecutiveIndicatorsDashboard
      parameters:
        - name: period
          in: query
          required: false
          schema: { type: string, pattern: "^\\d{4}-Q[1-4]$" }
      responses:
        "200":
          description: Indicadores executivos consolidados para o tenant
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/jurisdictions/country-groups:
    get:
      summary: Consultar o catálogo de países por bloco gerencial
      tags: [Jurisdictions]
      operationId: listCountryGroups
      description: >-
        Retorna a relação documental anonimizada. Presença operacional permanece não confirmada;
        o bloco de Ilhas do Caribe depende da decisão D-01.
      responses:
        "200":
          description: Catálogo documental agrupado
          content:
            application/json:
              schema:
                type: object
                required:
                  [companyDisplayName, operationalPresenceConfirmed, data]
                properties:
                  companyDisplayName:
                    { type: string, const: Empresa Confidencial }
                  operationalPresenceConfirmed: { type: boolean, const: false }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/CountryGroup" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/jurisdictions:
    get:
      summary: Listar jurisdições visíveis
      tags: [Jurisdictions]
      operationId: listJurisdictions
      responses:
        "200":
          description: Jurisdições visíveis para o tenant e escopos do ator
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Jurisdiction"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Criar uma jurisdição
      tags: [Jurisdictions]
      operationId: createJurisdiction
      description: Requer função tax-admin ou country-manager.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJurisdiction"
      responses:
        "201":
          description: Jurisdição criada e auditada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jurisdiction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/jurisdictions/{id}:
    patch:
      summary: Atualizar uma jurisdição
      tags: [Jurisdictions]
      operationId: updateJurisdiction
      description: Requer função tax-admin ou country-manager.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJurisdiction"
      responses:
        "200":
          description: Jurisdição atualizada e auditada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jurisdiction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Em produção, token do provedor corporativo. Cabeçalhos sintéticos existem apenas localmente.
...
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  requestBodies:
    DomainRecord:
      required: true
      description: >-
        Payload específico do domínio, validado estritamente pela API. Campos legais devem
        incluir fonte e status de validação quando aplicável.
      content:
        application/json:
          schema:
            type: object
            minProperties: 1
            additionalProperties: true
  responses:
    DomainRecordList:
      description: Registros visíveis no tenant e nos escopos de país do ator
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data:
                type: array
                items: { $ref: "#/components/schemas/DomainRecord" }
    DomainRecordCreated:
      description: Registro criado e evento de auditoria encadeado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/DomainRecord" }
    BadRequest:
      description: Requisição inválida
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Identidade ausente ou inválida
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: Identidade válida sem permissão efetiva
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Recurso inexistente ou não visível no tenant
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: Conflito com o estado atual
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    DomainRecord:
      type: object
      required: [id, tenantId, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        legalValidationStatus:
          type: string
          enum: [PRELIMINARY, PENDING_LOCAL_COUNSEL, VALIDATED, SUPERSEDED]
      additionalProperties: true
...
    CreateJurisdiction:
      type: object
      additionalProperties: false
      required: [countryCode, name, managementBlock, jurisdictionType]
      properties:
        countryCode: { type: string, pattern: "^[A-Za-z]{2}$", examples: [BR] }
        name: { type: string, minLength: 2, maxLength: 120, examples: [Brasil] }
        managementBlock:
          type: string
          enum: [CENTRAL_AMERICA, CARIBBEAN_ISLANDS, SOUTH_AMERICA]
        jurisdictionType:
          type: string
          enum:
            [
              SOVEREIGN_STATE,
              NON_SOVEREIGN_TERRITORY,
              CONSTITUENT_COUNTRY,
              OVERSEAS_TERRITORY,
              OVERSEAS_REGION,
            ]
        sovereignAuthorityCode:
          type: string
          pattern: "^[A-Za-z]{2}$"
          description: Obrigatório para jurisdições não soberanas e proibido para SOVEREIGN_STATE.
    SovereignAuthority:
      type: object
      additionalProperties: false
      required: [countryCode, name]
      properties:
        countryCode: { type: string, pattern: "^[A-Z]{2}$" }
        name: { type: string }
    JurisdictionCatalogItem:
      type: object
      additionalProperties: false
      required:
        [
          countryCode,
          name,
          jurisdictionType,
          sovereignAuthority,
          applicableLawModel,
          legalValidationStatus,
          tierReference,
          documentaryBasis,
          operationalPresence,
        ]
      properties:
        countryCode: { type: string, pattern: "^[A-Z]{2}$" }
        name: { type: string }
        jurisdictionType:
          type: string
          enum:
            [
              SOVEREIGN_STATE,
              NON_SOVEREIGN_TERRITORY,
              CONSTITUENT_COUNTRY,
              OVERSEAS_TERRITORY,
              OVERSEAS_REGION,
            ]
        sovereignAuthority:
          oneOf:
            - $ref: "#/components/schemas/SovereignAuthority"
            - type: "null"
        applicableLawModel:
          type: string
          enum:
            [
              SOVEREIGN_DOMESTIC_LAW,
              TERRITORY_SPECIFIC_WITH_SOVEREIGN_FRAMEWORK,
            ]
        legalValidationStatus: { type: string, const: REQUIRED_LOCAL_COUNSEL }
        legalFrameworkNote: { type: string }
        tierReference:
          type: [integer, "null"]
          minimum: 1
          maximum: 5
          description: Nulo quando o país consta apenas do catálogo regional e ainda não possui tier documental.
        documentaryBasis:
          type: string
          enum: [EXPLICITLY_NAMED, AGGREGATE_REFERENCE, REGIONAL_CATALOG]
        operationalPresence: { type: string, const: UNCONFIRMED }
    CountryGroup:
      type: object
      additionalProperties: false
      required: [code, name, coverageStatus, countries, territories]
      properties:
        code:
          type: string
          enum: [CENTRAL_AMERICA, CARIBBEAN_ISLANDS, SOUTH_AMERICA]
        name: { type: string }
        coverageStatus:
          type: string
          enum: [DOCUMENTED_REFERENCE, PENDING_CONFIRMATION]
        classificationNote: { type: string }
        countries:
          type: array
          items: { $ref: "#/components/schemas/JurisdictionCatalogItem" }
        territories:
          type: array
          items: { $ref: "#/components/schemas/JurisdictionCatalogItem" }
    UpdateJurisdiction:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        status: { type: string, enum: [active, inactive] }
    Jurisdiction:
      allOf:
        - $ref: "#/components/schemas/CreateJurisdiction"
        - type: object
          required:
            [
              id,
              tenantId,
              applicableLawModel,
              legalValidationStatus,
              status,
              createdAt,
              updatedAt,
            ]
          properties:
            id: { type: string, format: uuid }
            tenantId: { type: string }
            applicableLawModel:
              type: string
              enum:
                [
                  SOVEREIGN_DOMESTIC_LAW,
                  TERRITORY_SPECIFIC_WITH_SOVEREIGN_FRAMEWORK,
                ]
            legalValidationStatus:
              { type: string, const: REQUIRED_LOCAL_COUNSEL }
            status: { type: string, enum: [active, inactive] }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...
    Problem:
      type: object
      required: [type, title, status, detail, instance, correlationId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        correlationId: { type: string }
security:
  - bearerAuth: []